Prepare the branch for merge with master.

+++ include/thread.h
@@ -56,12 +56,14 @@ typedef struct thread {
   /* scheduler part */
   td_prio_t td_prio;
   int td_slice;
+  /* thread statistics */
   timeval_t td_rtime;        /* ticks spent running */
   timeval_t td_last_rtime;   /* time of last switch to running state */
   timeval_t td_slptime;      /* ticks spent sleeping */
   timeval_t td_last_slptime; /* time of last switch to sleep state */
-  uint64_t td_nctxsw;        /* total number of context switches */
-  sigset_t td_sigpend;       /* Pending signals for this thread. */
+  unsigned td_nctxsw;        /* total number of context switches */
+  /* signal handling */
+  sigset_t td_sigpend; /* Pending signals for this thread. */
   /* TODO: Signal mask, sigsuspend. */
 } thread_t;
